Add init container in CSI controller
The init container makes sure that /var/run/secrets/aws/credentials file exists
and contains the AWS credentials. This is useful when CCO is in manual mode and
doesn't create the aws-efs-cloud-credentials secret with the "credentials" key
(this is supposed to be created manually). This key is projected as a file
in the csi-driver container.